Store incoming information on eaters
Incoming chunk information is stored in a hashmap, along with
information on who has sent it. If information is saved about the same
path multiple times it means that there are multiple chunks, so we store
up to 16 sources along with the file information.

This hash map then basically contains out worklist, we can fill in some
appropriate P/Q values and broadcast it.

Currently all keys get a separate heap allocation, that the hash map
then frees if the key is already present. This is not very nice in terms
of interface, performance and memory usage -- but it could be replaced
by a simple bump allocator if needed.