Broadcast worklists - first draft
I also switched the ordering, so our ranks are contiguous in phase 2.
The next two things that need to be done are to find out an appropriate
way of transmitting the filepaths, we can't just send an array of
pointers.
Secondly we also need to process the file info list to add P/Q and maybe
map locations from feeder-ranks to eater-ranks.