Broadcast worklists - second draft
In stead of heap allocating each string they are stored contiguously in
`flat_file_names`. Every incoming path is added at the end, the info is
stored in the hash map and if the path was already present we remove the
path from the `flat_file_names` again.

When we need to broadcast our worklist we now look up each path in the
hash, in the order they are stored, rather than collecting whatever
order the hash table had stored them in. This means we don't have to
repack all the paths.