Add actual parity calculation
A simple naive loop optimizes better than the current solution, but when
compiling in debug mode the uint64 version is much faster than a naive
version.
Another possibility would be to have the xor_parity function in a
separate .c/.o file that is always compiled with optimizations, but that
would complicate the makefile.