Reorder feeder/eater ranks for multi-machine runs
The mapping from mpi ranks to eaters/feeders only worked for a single
machine, or if nodes were specified carefully in the hostfile. Now we
assume pairs of ranks are on the same machine as hostfiles will usually
specify the nodes in sorted order.

Additionally support for a variable number of storage targets, up to
MAX_TARGETS, was added.