Add persistent records for chunks
Everyone involved in processing a task stores the full task info to a
leveldb database that persist between runs. This means that if a single
node reports an update to a chunk the coordinator can fill in other
relevant locations, etc.

Currently it probably wont work if the total file size changes, we don't
really have a way of knowing it without talking to the other chunk
holders.