Change to more compact representation of locations
Instead of an unordered array of integers we now use a fixed size
56 bit bit-vector plus 8 bits to indicate P -- going from 32 bytes to 8
bytes per task.
This means we are limited in the overall number of storage targets
rather than in the number of chunks a file can be split into.

There is currently an unfortunate leftover in that the locations are
actually mpi ranks. It would be better to use storage target ids that
are guaranteed to not change between runs, ranks will probably change if
a target dies and we need to rebuild. Also it means that we only ever
use every second bit in the bit-vector meaning that we are currently
limited to 28 storage targets.