Give the filesystem a couple of hints
We use posix_fadvise to the kernel/filesystem that we are going to read
our input chunks sequentially and we are going to read all of their
contents. We also use posix_fallocate to indicate the expected size of
our parity chunk file.

We can't really test performance on the current test setup, but it seems
unlikely that this would hurt and it might give us a little extra speed.