Allow storage targets to be moved between hosts
If we have had failures and rebuilt a target it might not be on the same
host as it used to. That gives us a problem since mpi ranks are given by
host order and not storage target order we need to create a mapping from
st -> rank and from rank -> st.

This is done by everyone sending their target ID and rank to the
coordinator, who can then create the mappings we need.  This will break
if we run on a new set of target IDs, so if we ever want to expand
/store01 or /store02 rather than create a new /store03 we will probably
have to wipe all parity data.