Add control of chunks/parity read/write locations
To control where a parity chunk is written you pass a substitution
pattern to the task processing. It works by copying from your pattern
whenever there is a non-space character, and from the original
otherwise.
So "/store01/chunks/A/B/C" and the pattern "         parity" maps to
"/store01/parity/A/B/C".

This is needed later if we want to use the same task processing when
rebuilding - some ranks will have to load from parity/ instead of
chunks/ and the rebuild target should write to chunks/ rather than
parity/.