Add iteration to persistent db
In order to rebuild we need to be able to iterate over all the stored
key-value pairs. I currently make sure each key is zero terminated. I am
not sure it is still necessary and it might be a better to solution to
just make sure we store only zero terminated keys.