Add the actual rebuilding
Before we only stored the total size of all the chunks, now we store
the sizes of all chunks. This is used to find the final file size on the
rebuild target.

The task processing code can't completely ignore if it is rebuilding or
not.  When rebuilding we don't want to store the chunk sizes, we want to
receive them from the storage target that is actually reading from the
parity file. This means we have to pass in some extra info which means
we could probably drop the load/save patterns again.

When rebuilding we don't have access to the database on the rebuild
target so a "helper" is chosen that sends all items of interest to the
rebuild target. Another solution would be to give the rebuild target a
copy of the database before starting.