Stop spending 90% of the time clearing buffers
For the test case I am using there are tons of small files (4K or less).
But for everyone of them we cleared multiple 10MiB buffers. If all our
files are large it wouldn't make a difference, but for small files it
means spending most of the time doing nothing but clearing buffers.