Remove reliance on max_chunk_size in process_task
We are going to have to stat the files to compare timestamps anyways, so
we might as well get the sizes our self.
A local coordinator receives chunk sizes from all chunk holders at the
start of the processing now, rather than at the end.
Then the coordinator send out the max chunk sizes to all the chunk
holders. This gives one synchronous message of overhead.

I also moved the chunk sizes to the start of the parity file since we
have them right away and we need them first thing when doing a rebuild.

I have not removed the max_chunk_size field completely as we are
currently using it to drive the progress reporting. I guess we need to
change the process_task interface to get the information out.