Track and print serious errors
The only IO error we can really accept is that chunks have gone away. It
just means that there will be a future unlink event.
If we get other errors, like disk failures, devices running out of space
or similar, we are going to assume that everything is broken. So as soon
as a host sees an error, it will write to /dev/null and read from
/dev/zero for the rest of the run.
After an error like this, the databases and parity data should be
cleared, the problem fixed and a complete parity generation run.