Remove POSIX_FADV_WILLNEED hint
The hint basically causes linux to read as much as possible of the file
into the file cache. Which wouldn't be too bad, except that it blocks
while doing this so there are long pauses before we can start processing
larger files.
The sequential hint just increases the read-ahead buffer size a bit, so
I've left it in place even though I haven't actually compared
performance with and without it.