Use the database instead of a helper for rebuilds
We can just copy the database from one of the hosts that are still in a
good state, and then use the database on all storage targets.
This way we don't need a 'helper' to send extra messages and it will
probably be easier to make the rebuild process multithreaded if
necessary.