Check for version mismatch
To prevent running with a different version on some of the storage
targets we make sure that everything is compiled from the same git
commit. The alternative would be to check either a __TIMESTAMP__ or a
hash of the actual executable.
That would either require the executable to be built on a single machine
and copied to everywhere else or the exact same compiler/library
versions on all storage targets.