Check which nodes are actually in use
Before you had to manually update your hostlist after migrating to a
spare node. Now the list is all potential machines, including spares.
We actually connect and check if they have the targetNumID file in the
currently processing store folder.

Also now the nodes are actually in a random order which means I've
noticed that the RunData struct wasn't being saved when generating
parity - which means the rebuild failed.