Rebuild target is unknown, choose target id
The rebuild target was the index in the host list of the host to
rebuild. Since the host list might be in a different order between runs,
we cannot depend on the index. We give the targetNumID instead, which
the program can then use to find its own index in the list of targets.