Simplify changelogger
By extracting which store we are working on from the parameter-list in
the wrapper we can get rid of the complicated init_once function.

Using syslog for logging instead of an open/write/close cycle simplifies
logging and lets us remove the last lock.

The commit also includes some renaming, reordering, reformatting and a
few comments.