Avoid zero weight storage targets
If all the possible places a parity block could be placed had a weight
of zero, we would loop forever trying to select one.
This change makes sure weights are calculated in a way that always give
a positive value. We might have to try a lot of values before we hit the
available low-weight target, but the random generator should be pretty
fast and it is hopefully a rare case.