Check for free_space.override in store dir
This is a file that can be used to pretend that a storage target has
less space than it really does. It can be used for testing or to force a
more aggresive balancing.
The code doesn't attempt any clever parsing, it just assumes there is
nothing but a base 10 number in the file.