Change how changelogs are read and deleted
In stead of keeping track of timestamps we will now read _all_
changelogs, and keep track of which of the ones we read that did not
have a lock. If there is no lock it means the changelogger has moved on
to an new file and we can safely remove the file.

This means we will end up processing some chunks multiple times, but we
can update the C code to skip work items that haven't changed relative
to what is stored in the leveldb database.